# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SupportDesk API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, in-progress, resolved]
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [bug, feature, billing, account, other]
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        "200":
          description: Tickets fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketsResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
  /tickets/satisfaction:
    get:
      operationId: getTicketSatisfaction
      tags: [tickets]
      security:
        - bearerAuth: []
      summary: Customer satisfaction summary
      responses:
        "200":
          description: Satisfaction summary fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSatisfactionResponse"
...
  /tickets/{ticketId}:
    get:
      operationId: getUserTicketDetail
      tags: [tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket detail with replies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTicketDetailResponse"
  /tickets/{ticketId}/replies:
    post:
      operationId: createUserTicketReply
      tags: [tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Reply posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
  /tickets/{ticketId}/rating:
    patch:
      operationId: rateTicket
      tags: [tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateTicketRequest"
      responses:
        "200":
          description: Ticket rated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTicketDetailResponse"
  /tickets/{ticketId}/attachments:
    get:
      operationId: getTicketAttachments
      tags: [tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attachments fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentsResponse"
    post:
      operationId: uploadTicketAttachment
      tags: [tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadAttachmentRequest"
      responses:
        "201":
          description: Attachment uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        productName:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in-progress, resolved]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
          enum: [bug, feature, billing, account, other]
        assignedTo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        feedbackText:
          type: string
          nullable: true
        slaDueAt:
          type: string
          format: date-time
          nullable: true
        slaEscalated:
          type: boolean
      required: [id, userId, productName, description, status, priority, category, createdAt, slaEscalated]
...
    TicketReply:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
          enum: [admin, user]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, ticketId, authorId, authorName, authorRole, message, createdAt]
    Attachment:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        uploadedBy:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        url:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, ticketId, uploadedBy, fileName, mimeType, fileSize, url, createdAt]
...
    UserTicketDetail:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        productName:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in-progress, resolved]
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
          enum: [bug, feature, billing, account, other]
        assignedTo:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        replies:
          type: array
          items:
            $ref: "#/components/schemas/TicketReply"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        rating:
          type: integer
          nullable: true
        feedbackText:
          type: string
          nullable: true
        slaDueAt:
          type: string
          format: date-time
          nullable: true
        slaEscalated:
          type: boolean
      required: [id, userId, productName, description, status, priority, category, createdAt, replies, attachments, slaEscalated]
...
    CreateTicketRequest:
      type: object
      properties:
        productName:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
          enum: [bug, feature, billing, account, other]
      required: [productName, description]
...
    CreateReplyRequest:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RateTicketRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedbackText:
          type: string
          nullable: true
      required: [rating]
    UploadAttachmentRequest:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        base64Data:
          type: string
      required: [fileName, mimeType, fileSize, base64Data]
...
    TicketResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/Ticket"
      required: [success, message, data]
    TicketsResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        meta:
          $ref: "#/components/schemas/TicketSearchMeta"
      required: [success, message, data, meta]
    TicketSearchMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, pageSize, total, totalPages]
    RatingDistributionItem:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        count:
          type: integer
      required: [rating, count]
    SatisfactionFeedback:
      type: object
      properties:
        ticketId:
          type: string
        productName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedbackText:
          type: string
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [ticketId, productName, rating, feedbackText, resolvedAt, createdAt]
...
    UserSatisfaction:
      type: object
      properties:
        ratedTicketCount:
          type: integer
        averageRating:
          type: number
          nullable: true
        distribution:
          type: array
          items:
            $ref: "#/components/schemas/RatingDistributionItem"
        recentFeedback:
          type: array
          items:
            $ref: "#/components/schemas/SatisfactionFeedback"
      required: [ratedTicketCount, averageRating, distribution, recentFeedback]
    UserSatisfactionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserSatisfaction"
      required: [success, message, data]
...
    UserTicketDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserTicketDetail"
      required: [success, message, data]
    AttachmentsResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
      required: [success, message, data]
    AttachmentResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/Attachment"
      required: [success, message, data]
...
    ReplyResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/TicketReply"
      required: [success, message, data]
...